Micro refactoring: move replanting methods to evaluator
