+++ envergo/hedges/services.py
@@ -1,3 +1,4 @@
+from collections import defaultdict
 from dataclasses import dataclass
 from enum import Enum
 from itertools import product
@@ -65,6 +66,9 @@ def _check_plantation_result_matrix():
 _check_plantation_result_matrix()
 
 
+R = 2
+
+
 @dataclass
 class EvaluationResult:
     result: Literal[PlantationResults.Adequate, PlantationResults.Inadequate]
@@ -125,6 +129,36 @@ class PlantationEvaluator:
             raise RuntimeError("Call the evaluator `evaluate` method first")
 
         return self._evaluation_result.conditions
+
+    def minimum_length_to_plant(self):
+        """Returns the minimum length of hedges to plant, considering the length of hedges to remove and the
+        replantation coefficient"""
+        return R * self.hedge_data.length_to_remove()
+
+    def get_minimum_lengths_to_plant(self):
+        lengths_by_type = defaultdict(int)
+        for to_remove in self.hedge_data.hedges_to_remove():
+            lengths_by_type[to_remove.hedge_type] += to_remove.length
+
+        return {
+            "degradee": R * lengths_by_type["degradee"],
+            "buissonnante": R * lengths_by_type["buissonnante"],
+            "arbustive": R * lengths_by_type["arbustive"],
+            "mixte": R * lengths_by_type["mixte"],
+            "alignement": R * lengths_by_type["alignement"],
+        }
+
+    def get_lengths_to_plant(self):
+        lengths_by_type = defaultdict(int)
+        for to_plant in self.hedge_data.hedges_to_plant():
+            lengths_by_type[to_plant.hedge_type] += to_plant.length
+
+        return {
+            "buissonnante": lengths_by_type["buissonnante"],
+            "arbustive": lengths_by_type["arbustive"],
+            "mixte": lengths_by_type["mixte"],
+            "alignement": lengths_by_type["alignement"],
+        }
 
     @property
     def evaluation(self):
